chatbot_backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic
langchain
langchain-openai
